requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
ddt==1.2.1
idna==2.8
Logbook==1.4.3
paramiko==2.4.2
pexpect==4.7.0
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
